Refaktorering av kode
Refaktorering av kode for å separare/skille ut kode for å gjøre komponenter mindre.

Separerte ut visning av knappene/våpnene for å gjøre komponenten mindre.
